Remove the *default-handler* var, and its concept
The *default-handler* was only used when no other handler had been
registered. The problem with that was that it didn't have its own
try/catch wrapper, and so it always behaved a bit weird. Rather than
keep putting effort into making it work, just remove it. If we find we
need it later we can always add it back in.