Remove special binding handling for handlers
I've decided this isn't a good idea. The idea was to make it so that
conceptually the handlers were executing higher up in the stack (like
a temporary stack unwind). This makes some sense if we consider this
through the continuation lens, but I think this ends up being more
work for little reward. This also means that there is some dynamic
state that the handlers are unable to access, which might reduce their
utility.

It's always possible for a user to manipulate the bindings themselves
if they want the old behaviour, but if we manipulate the bindings then
it's hard for a user to access the proper dynamic scope.