Add a type parameter to prompts, for forms, files, etc.
For instance, you can use :form to indicate that the user's
environment should provide syntax help to them entering a form. Or you
can use :file to indicate the user's environment should prompt them
for a file specifically.

There is no defined set of "correct" keys, so the nrepl support will
have to handle things that it can, and default to a plain string if it
can't understand what it's being asked to do.