nrepl: Make execution sequential
We do this using locks, making sure that only one thing is executing
at a time. We also then use the locks to signal between threads when
there might be data for them to consume.

It's all a little bit hacky at the moment, but I think it will work.