elisp: Still prompt, even when no restarts are available
This is necessary because there are two "default" restarts, that are
always valid. No matter what the exception is, we always have the
option to leave it unhandled, or to abort the computation. The
previous check was too strict, and would sometimes result in the
thread being stuck with no way to respond and abort it.